Refactor validator
