+++ operator/controllers/mlops/seldonconfig_controller.go
@@ -12,15 +12,12 @@ package mlops
 import (
 	"context"
 
-	"github.com/go-logr/logr"
-	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	mlopsv1alpha1 "github.com/seldonio/seldon-core/operator/v2/apis/mlops/v1alpha1"
-	"github.com/seldonio/seldon-core/operator/v2/controllers/reconcilers/common"
 )
 
 // SeldonConfigReconciler reconciles a SeldonConfig object
@@ -43,25 +40,11 @@ type SeldonConfigReconciler struct {
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.13.0/pkg/reconcile
 func (r *SeldonConfigReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	logger := log.FromContext(ctx).WithName("Reconcile")
+	_ = log.FromContext(ctx)
 
-	// Fetch the SeldonConfig instance
-	seldonConfig := &mlopsv1alpha1.SeldonConfig{}
-	if err := r.Client.Get(ctx, req.NamespacedName, seldonConfig); err != nil {
-		if errors.IsNotFound(err) {
-			logger.Error(err, "unable to fetch SeldonConfig", "name", req.Name, "namespace", req.Namespace)
-			return ctrl.Result{}, err
-		}
-		return ctrl.Result{}, nil
-	}
+	// TODO(user): your logic here
 
-	err := ValidateSeldonConfig(
-		ctx, r.Client,
-		seldonConfig,
-		nil,
-		logr.FromContextOrDiscard(ctx),
-	)
-	return ctrl.Result{}, err
+	return ctrl.Result{}, nil
 }
 
 // SetupWithManager sets up the controller with the Manager.
@@ -69,28 +52,4 @@ func (r *SeldonConfigReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&mlopsv1alpha1.SeldonConfig{}).
 		Complete(r)
-}
-
-func ValidateSeldonConfig(
-	ctx context.Context,
-	clt client.Client,
-	seldonConfig *mlopsv1alpha1.SeldonConfig,
-	namespace *string,
-	logger logr.Logger,
-) error {
-	for _, component := range seldonConfig.Spec.Components {
-		logger.Info("Checking component", "component_name", component.Name)
-
-		if component.Name == mlopsv1alpha1.DataflowEngineName {
-			return common.ValidateDataflowScaleSpec(
-				ctx,
-				clt,
-				component,
-				&seldonConfig.Spec.Config.KafkaConfig,
-				namespace,
-				logger,
-			)
-		}
-	}
-	return nil
 }
